requirements-dev.txt
ruff==0.7.4
black==24.10.0

mysql-connector-python==9.1.0
